Do all the chart sizing work in same batch
